Renamed extension function to mapTo and moved to web package
